Store files in database

+++ files/delete.php
@@ -10,8 +10,8 @@ if(isset($_POST['delete_btn'])) {
     $file_server = $file_path_server . $_POST['delete_btn'];
 
     try {
-        $stmt = $pdo-> prepare("DELETE FROM `files` WHERE `filename` = ?;");
-        $stmt-> execute([$file_name]);
+        $stmt = $pdo-> prepare("DELETE FROM `files` WHERE `user_id` = ? AND `filename` = ?;");
+        $stmt-> execute([$_SESSION['user'], $file_name]);
     } catch (\PDOException $e) {
         throw new \PDOException($e-> getMessage(), (int)$e-> getCode());
     }
